+++ operations/rra.c
@@ -6,16 +6,16 @@
 /*   By: jkauker <jkauker@student.42heilbrnn.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 16:08:53 by jkauker           #+#    #+#             */
-/*   Updated: 2023/12/14 13:45:04 by jkauker          ###   ########.fr       */
+/*   Updated: 2023/12/15 09:38:45 by jkauker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
 // rra : reverse rotate a - shift down all elements of stack a by 1. The last
-void	rra(int *stack_a, int *size, int print)
+void	rra(long *stack_a, int *size, int print)
 {
-	int		tmp;
+	long	tmp;
 	int		i;
 
 	tmp = stack_a[*size - 1];
