+++ operations/sa.c
@@ -6,16 +6,16 @@
 /*   By: jkauker <jkauker@student.42heilbrnn.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 16:06:50 by jkauker           #+#    #+#             */
-/*   Updated: 2023/12/14 13:47:57 by jkauker          ###   ########.fr       */
+/*   Updated: 2023/12/15 09:38:59 by jkauker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
 // sawp a - swap the first 2 elements at the top of stack a. Do nothing if there
-void	sa(int *stack_a, int print)
+void	sa(long *stack_a, int print)
 {
-	int		tmp;
+	long	tmp;
 
 	tmp = stack_a[0];
 	stack_a[0] = stack_a[1];
