+++ operations/sb.c
@@ -6,16 +6,16 @@
 /*   By: jkauker <jkauker@student.42heilbrnn.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 16:07:10 by jkauker           #+#    #+#             */
-/*   Updated: 2023/12/14 13:48:17 by jkauker          ###   ########.fr       */
+/*   Updated: 2023/12/15 09:35:31 by jkauker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
 // sawp b - swap the first 2 elements at the top of stack b. Do nothing if there
-void	sb(int *stack_b, int print)
+void	sb(long *stack_b, int print)
 {
-	int		tmp;
+	long		tmp;
 
 	tmp = stack_b[0];
 	stack_b[0] = stack_b[1];
