+++ operations/ss.c
@@ -6,14 +6,14 @@
 /*   By: jkauker <jkauker@student.42heilbrnn.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/29 16:07:26 by jkauker           #+#    #+#             */
-/*   Updated: 2023/12/14 13:48:33 by jkauker          ###   ########.fr       */
+/*   Updated: 2023/12/15 09:38:56 by jkauker          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
 // ss : sa and sb at the same time.
-void	ss(int *stack_a, int *stack_b)
+void	ss(long *stack_a, long *stack_b)
 {
 	sa(stack_a, 0);
 	sb(stack_b, 0);
